Prep the GitHub Actions setup for handling more than one E2E test (#106)
